+++ ElaWidgetTools/DeveloperComponents/ElaMenuBarStyle.cpp
@@ -55,7 +55,7 @@ void ElaMenuBarStyle::drawControl(ControlElement element, const QStyleOption* op
             QFont iconFont = QFont("ElaAwesome");
             iconFont.setPixelSize(18);
             painter->setFont(iconFont);
-            painter->drawText(topt->rect, Qt::AlignCenter, QChar((unsigned short)ElaIconType::AngleRight));
+            painter->drawText(topt->rect, Qt::AlignCenter, QChar(ElaIconType::AngleRight));
             painter->restore();
         }
         return;
