+++ ElaWidgetTools/ElaIcon.cpp
@@ -23,7 +23,7 @@ QIcon ElaIcon::getElaIcon(ElaIconType::IconName awesome)
     // painter.setBrush(QColor("#1570A5"));
     iconFont.setPixelSize(25);
     painter.setFont(iconFont);
-    painter.drawText(pix.rect(), Qt::AlignCenter, QChar((unsigned short)awesome));
+    painter.drawText(pix.rect(), Qt::AlignCenter, QChar(awesome));
     painter.end();
     return QIcon(pix);
 }
@@ -40,7 +40,7 @@ QIcon ElaIcon::getElaIcon(ElaIconType::IconName awesome, QColor iconColor)
     // painter.setBrush(QColor("#1570A5"));
     iconFont.setPixelSize(25);
     painter.setFont(iconFont);
-    painter.drawText(pix.rect(), Qt::AlignCenter, QChar((unsigned short)awesome));
+    painter.drawText(pix.rect(), Qt::AlignCenter, QChar(awesome));
     painter.end();
     return QIcon(pix);
 }
@@ -56,7 +56,7 @@ QIcon ElaIcon::getElaIcon(ElaIconType::IconName awesome, int pixelSize)
     iconFont.setPixelSize(pixelSize);
     painter.setFont(iconFont);
     // 画图形字体
-    painter.drawText(pix.rect(), Qt::AlignCenter, QChar((unsigned short)awesome));
+    painter.drawText(pix.rect(), Qt::AlignCenter, QChar(awesome));
     painter.end();
     return QIcon(pix);
 }
@@ -73,7 +73,7 @@ QIcon ElaIcon::getElaIcon(ElaIconType::IconName awesome, int pixelSize, QColor i
     iconFont.setPixelSize(pixelSize);
     painter.setFont(iconFont);
     // 画图形字体
-    painter.drawText(pix.rect(), Qt::AlignCenter, QChar((unsigned short)awesome));
+    painter.drawText(pix.rect(), Qt::AlignCenter, QChar(awesome));
     painter.end();
     return QIcon(pix);
 }
@@ -89,7 +89,7 @@ QIcon ElaIcon::getElaIcon(ElaIconType::IconName awesome, int pixelSize, int fixe
     iconFont.setPixelSize(pixelSize);
     painter.setFont(iconFont);
     // 画图形字体
-    painter.drawText(pix.rect(), Qt::AlignCenter, QChar((unsigned short)awesome));
+    painter.drawText(pix.rect(), Qt::AlignCenter, QChar(awesome));
     painter.end();
     return QIcon(pix);
 }
@@ -106,7 +106,7 @@ QIcon ElaIcon::getElaIcon(ElaIconType::IconName awesome, int pixelSize, int fixe
     iconFont.setPixelSize(pixelSize);
     painter.setFont(iconFont);
     // 画图形字体
-    painter.drawText(pix.rect(), Qt::AlignCenter, QChar((unsigned short)awesome));
+    painter.drawText(pix.rect(), Qt::AlignCenter, QChar(awesome));
     painter.end();
     return QIcon(pix);
 }
